Signed-off-by: jingfu wang <jingfu.j.wang@ibm.com>

**What type of PR is this?**
/kind feature
/area devfile

**What does does this PR do / why we need it**:
This PR implements the `--devfile` flag support with the following details:

1. Features
- [x] When `--devfile` is set, odo should download/copy devfile from the specific location to user's working directory
- [x] The value of `--devfile` should support both http(s) path and file system path
- [x] Remove `--devfile` for all other commands (eg. `odo url`, `odo push`, `odo delete`, `odo watch`), which means `--devfile` only be used by `odo create` and all other commands can directly use the devile in user's working directory instead of specifying the devfile path separately
- [x] New design of existing devfile support: if user decides to use existing devfile (devfile exists in user's working directory or user specifies the existing devfile by `--devfile`) the first argument will be the component name instead of component type, and user can pass in the component name directly in direct mode rather than interactive mode

2. Handling errors
- [x] Validate existence and correctness of `--devfile` value
- [x] If a devfile already exists in the user's working directory, error out if user still specify **another** existing devfile by `--devfile`
- [x] `--devfile` and `--registry` should be mutually exclusive flags

**Which issue(s) this PR fixes**:

Fixes #2862 

**How to test changes / Special notes to the reviewer**:
Run 
- `odo create [component name] --devfile <your own devfile path>`