@dharmit 2020-05-08
Should this validation not happen before we ask the user for the details? I think this is because of the `co.Interactive` in `Complete` function while the validation happening in `Validate`.

If it makes sense, we can create an issue out of it and discuss it there.

```sh
$ odo create --devfile /tmp/devfile.yaml
Experimental mode is enabled, use at your own risk

? What do you wish to name the new devfile component nothing
? What namespace do you want the devfile component to be created in newproject
 ✗  Validating devfile component [69997ns]
 ✗  This workspace directory already contains a devfile component
```