@dharmit 2020-05-08
This one's got me confused. Why does it take us 9 seconds to tell the user that a `devfile.yaml` already exists in `pwd`? Shouldn't that be faster? :thinking: 
```sh
$ time odo create --devfile /tmp/devfile.yaml
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a devfile.yaml, you can't specify devfile path via --devfile
odo create --devfile /tmp/devfile.yaml  0.15s user 0.02s system 1% cpu 9.276 total
```