@adisky 2020-05-08
> This one's got me confused. Why does it take us 9 seconds to tell the user that a `devfile.yaml` already exists in `pwd`? Shouldn't that be faster? 

I have raised this point earlier too, actually call to  `Listdevfilecomponents` is the cause of the delay, if possible all required error checks should be made before this call. https://github.com/openshift/odo/blob/9411894b1b50ea0caf7f09e7810848e6997030ef/pkg/odo/cli/component/create.go#L332