@dharmit 2020-05-08
> I have raised this point earlier too, actually call to `Listdevfilecomponents` is the cause of the delay, if possible all required error checks should be made before this call.

In that case, I guess #3112 should help address the concerns in terms of seconds. But yes I do feel like some checks should be made sooner rather than later.