@GeekArthur 2020-05-08
> OK, this gets interesting. Continuing from my above comment:
> 
> ```shell
> $ odo create --devfile /tmp/devfile.yaml
> Experimental mode is enabled, use at your own risk
> 
> ? What do you wish to name the new devfile component nothing
> ? What namespace do you want the devfile component to be created in newproject
>  ✗  Validating devfile component [84901ns]
>  ✗  This workspace directory already contains a devfile component
> 
> $ odo delete --all -f
>  ✗  The current directory does not represent an odo component. Use 'odo create' to create component here or switch to directory with a component
> 
> $ tree .odo
> .odo
> └── env
>     └── env.yaml
> 
> 1 directory, 1 file
> 
> $ ls devfile.yaml
> ls: cannot access 'devfile.yaml': No such file or directory
> ```
> 
> So, for `odo create`, existence of `env.yaml` is an indicator that `pwd` contains a component. But for `odo delete`, it isn't. Am I looking at this the right way or is my understanding flawed? @GeekArthur @johnmcollier @kadel
> 
> **EDIT**: I'm not suggesting we fix it in this PR. It's a separate issue and PR of it's own, provided my understanding is correct.

I think `env.yaml` is the only indicator that we can use to decide if the devfile component is present or not and `odo delete` should check that indicator and clean that up with user's confirmation. It's similar to s2i component, I think only difference is devfile component uses `env.yaml` and s2i component uses `config.yaml`.