@GeekArthur 2020-05-08
> This one's got me confused. Why does it take us 9 seconds to tell the user that a `devfile.yaml` already exists in `pwd`? Shouldn't that be faster? 🤔
> 
> ```shell
> $ time odo create --devfile /tmp/devfile.yaml
> Experimental mode is enabled, use at your own risk
> 
>  ✗  This directory already contains a devfile.yaml, you can't specify devfile path via --devfile
> odo create --devfile /tmp/devfile.yaml  0.15s user 0.02s system 1% cpu 9.276 total
> ```
> 
> **EDIT 1**: Same for specifying component type along with `--devfile` flag:
> 
> ```shell
> $ time odo create nodejs --devfile /tmp/devfile.yaml
> Experimental mode is enabled, use at your own risk
> 
>  ✗  Component type and --devfile can not be specified together
> odo create nodejs --devfile /tmp/devfile.yaml  0.18s user 0.01s system 1% cpu 14.590 total
> ```
> 
> I understand that the behaviour I've reported here and in previous comments doesn't indicate any bug. But this is sending a bad user experience, IMO.
> 
> **EDIT 2**: If I specify an invalid file to the `--devfile` flag, it's not throwing an error right away but taking 10+ seconds. See below:
> 
> ```shell
> $ time odo create nodejs --devfile /tmp/nosuchfile
> Experimental mode is enabled, use at your own risk
> 
>  ✗  Component type and --devfile can not be specified together
> odo create nodejs --devfile /tmp/nosuchfile  0.16s user 0.03s system 1% cpu 11.891 total
> 
> $ time odo create --devfile /tmp/nosuchfile      
> Experimental mode is enabled, use at your own risk
> 
> ? What do you wish to name the new devfile component nothing
> ? What namespace do you want the devfile component to be created in newproject
>  ✓  Validating devfile component [48332ns]
>  ✗  stat /tmp/nosuchfile: no such file or directory
> odo create --devfile /tmp/nosuchfile  0.18s user 0.03s system 1% cpu 15.457 total
> ```
> 
> I'm sorry but I'm going to hold this PR till the UX issues are addressed. @kadel @elsony @girishramnani @johnmcollier if you think these issues are OK to live with, feel free to remove the hold.
> 
> /hold

Thanks for testing those error handling! Although my personal system doesn't have the delay issue, I still need to fix this to tolerate all systems. The root cause of all scenarios can be boiled down to calling this function `Listdevfilecomponents()`, I can do some logic change to improve the performance to cover all scenarios that you test.