@GeekArthur 2020-05-08
> This one's got me confused. Why does it take us 9 seconds to tell the user that a `devfile.yaml` already exists in `pwd`? Shouldn't that be faster? 🤔
> 
> ```shell
> $ time odo create --devfile /tmp/devfile.yaml
> Experimental mode is enabled, use at your own risk
> 
>  ✗  This directory already contains a devfile.yaml, you can't specify devfile path via --devfile
> odo create --devfile /tmp/devfile.yaml  0.15s user 0.02s system 1% cpu 9.276 total
> ```
> 
> **EDIT 1**: Same for specifying component type along with `--devfile` flag:
> 
> ```shell
> $ time odo create nodejs --devfile /tmp/devfile.yaml
> Experimental mode is enabled, use at your own risk
> 
>  ✗  Component type and --devfile can not be specified together
> odo create nodejs --devfile /tmp/devfile.yaml  0.18s user 0.01s system 1% cpu 14.590 total
> ```
> 
> I understand that the behaviour I've reported here and in previous comments doesn't indicate any bug. But this is sending a bad user experience, IMO.
> 
> **EDIT 2**: If I specify an invalid file to the `--devfile` flag, it's not throwing an error right away but taking 10+ seconds. See below:
> 
> ```shell
> $ time odo create nodejs --devfile /tmp/nosuchfile
> Experimental mode is enabled, use at your own risk
> 
>  ✗  Component type and --devfile can not be specified together
> odo create nodejs --devfile /tmp/nosuchfile  0.16s user 0.03s system 1% cpu 11.891 total
> 
> $ time odo create --devfile /tmp/nosuchfile      
> Experimental mode is enabled, use at your own risk
> 
> ? What do you wish to name the new devfile component nothing
> ? What namespace do you want the devfile component to be created in newproject
>  ✓  Validating devfile component [48332ns]
>  ✗  stat /tmp/nosuchfile: no such file or directory
> odo create --devfile /tmp/nosuchfile  0.18s user 0.03s system 1% cpu 15.457 total
> ```
> 
> I'm sorry but I'm going to hold this PR till the UX issues are addressed. @kadel @elsony @girishramnani @johnmcollier if you think these issues are OK to live with, feel free to remove the hold.
> 
> /hold

@dharmit My latest commit improves the performance of error handling, you can check again on your system

Case 1: Component is present, **From ~9s to ~0.6s**
```
time ( repeat 10; odo create --devfile /Users/jingfuwang/wang/devfile.yaml)                                  ✔  3s   17:32:49 
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
Experimental mode is enabled, use at your own risk

 ✗  This directory already contains a component
( repeat 10; do; odo create --devfile /Users/jingfuwang/wang/devfile.yaml; ; )  0.69s user 0.16s system 128% cpu 0.652 total
```

Case 2: Component type conflict, **From ~15s to ~0.6**
```
time ( repeat 10; odo create nodejs --devfile /fakefile)                                                           1 ✘  17:31:59 
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
Experimental mode is enabled, use at your own risk

 ✗  Component type and --devfile can not be specified together
( repeat 10; do; odo create nodejs --devfile /fakefile; done; )  0.73s user 0.18s system 131% cpu 0.690 total
```

Case 3: Fake devfile path, **From ~15s to 0.6s**
```
time ( repeat 10; odo create --devfile /fakefile)                                                                  1 ✘  17:32:26 
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
Experimental mode is enabled, use at your own risk

 ✗  stat /fakefile: no such file or directory
( repeat 10; do; odo create --devfile /fakefile; done; )  0.72s user 0.16s system 130% cpu 0.675 total
```